move max width to the right place
