+++ src/app/component/[component]/layout.tsx
@@ -1,22 +1,22 @@
-"use client";
+'use client';
 
-import { css } from "@emotion/css";
-import { useRouter, usePathname } from "next/navigation";
-import React from "react";
+import { css } from '@emotion/css';
+import { useRouter, usePathname } from 'next/navigation';
+import React from 'react';
 
-import IconButton from "@leafygreen-ui/icon-button";
-import { CodeSandbox, Figma, Github } from "@/components/glyphs";
-import { useDarkMode } from "@leafygreen-ui/leafygreen-provider";
-import { Tabs, Tab } from "@leafygreen-ui/tabs";
-import { color, spacing } from "@leafygreen-ui/tokens";
-import { H2 } from "@leafygreen-ui/typography";
+import IconButton from '@leafygreen-ui/icon-button';
+import { CodeSandbox, Figma, Github } from '@/components/glyphs';
+import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
+import { Tabs, Tab } from '@leafygreen-ui/tabs';
+import { color, spacing } from '@leafygreen-ui/tokens';
+import { H2 } from '@leafygreen-ui/typography';
 
-import useComponentFields from "@/hooks/useComponentFields";
-import { getGithubLink } from "@/utils";
+import { useComponentFields } from '@/hooks';
+import { getGithubLink } from '@/utils';
 
-const liveExamplePath = "live-example";
-const designDocsPath = "design-docs";
-const codeDocsPath = "code-docs";
+const liveExamplePath = 'live-example';
+const designDocsPath = 'design-docs';
+const codeDocsPath = 'code-docs';
 
 export default function ComponentLayout({
   children,
@@ -25,13 +25,13 @@ export default function ComponentLayout({
 }) {
   const router = useRouter();
   const pathname = usePathname();
-  const currentComponent = pathname.split("/")[2];
+  const currentComponent = pathname.split('/')[2];
   const { theme } = useDarkMode();
 
   const component = useComponentFields({ componentName: currentComponent });
 
   const getSelected = () => {
-    const suffix = pathname.split("/")[3];
+    const suffix = pathname.split('/')[3];
     if (suffix === liveExamplePath) {
       return 0;
     }
@@ -61,7 +61,7 @@ export default function ComponentLayout({
       href: component?.codesandbox_url?.href,
       icon: <CodeSandbox />,
     },
-  ]
+  ];
 
   return (
     <div
@@ -75,7 +75,7 @@ export default function ComponentLayout({
           margin-bottom: ${spacing[600]}px;
         `}
       >
-        {currentComponent.split("-").join(" ")}
+        {currentComponent.split('-').join(' ')}
       </H2>
       <Tabs
         selected={getSelected()}
@@ -94,18 +94,20 @@ export default function ComponentLayout({
               height: 100%;
             `}
           >
-            {externalLinks.map(({ 'aria-label': ariaLabel, href, icon }, index) => (
-              <IconButton
-                key={ariaLabel + index}
-                aria-label={ariaLabel}
-                size="large"
-                href={href}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                {icon}
-              </IconButton>
-            ))}
+            {externalLinks.map(
+              ({ 'aria-label': ariaLabel, href, icon }, index) => (
+                <IconButton
+                  key={ariaLabel + index}
+                  aria-label={ariaLabel}
+                  size="large"
+                  href={href}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                >
+                  {icon}
+                </IconButton>
+              ),
+            )}
           </div>
         }
       >
