+++ src/components/global/SideNavigation/SideNavigation.tsx
@@ -22,7 +22,7 @@ import {
 import { MongoDBLogo, SupportedColors } from '@leafygreen-ui/logo';
 import { color, spacing } from '@leafygreen-ui/tokens';
 import { SIDE_NAV_WIDTH } from '@/constants';
-import { useMediaQuery } from '@/hooks/useMediaQuery';
+import { useMediaQuery } from '@/hooks';
 import { ComponentMeta, Group, groupedComponents } from '@/utils/components';
 import { Search } from '../Search';
 import { Drawer } from './Drawer';
