+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useCallbackRef } from './useCallbackRef';
-export { default as useComponentFields } from './useComponentFields';
+export { useComponentFields } from './useComponentFields';
 export { useMediaQuery } from './useMediaQuery';
 export { useSession } from './useSession';
